// src/api/models/GameRecordOutput.ts
/* tslint:disable */
/* eslint-disable */
/**
 * radarvan
 * Stats for generals
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { MatchListing } from './MatchListing';
import {
    MatchListingFromJSON,
    MatchListingFromJSONTyped,
    MatchListingToJSON,
    MatchListingToJSONTyped,
} from './MatchListing';

/**
 * 
 * @export
 * @interface GameRecordOutput
 */
export interface GameRecordOutput {
    /**
     * 
     * @type {string}
     * @memberof GameRecordOutput
     */
    jsonS3Uri: string;
    /**
     * 
     * @type {number}
     * @memberof GameRecordOutput
     */
    fileSizeBytes?: number | null;
    /**
     * 
     * @type {Date}
     * @memberof GameRecordOutput
     */
    gameTimestamp: Date;
    /**
     * 
     * @type {number}
     * @memberof GameRecordOutput
     */
    matchId: number;
    /**
     * 
     * @type {string}
     * @memberof GameRecordOutput
     */
    replayFileUrl: string;
    /**
     * 
     * @type {Date}
     * @memberof GameRecordOutput
     */
    createdAt: Date;
    /**
     * 
     * @type {Date}
     * @memberof GameRecordOutput
     */
    gameDate: Date;
    /**
     * 
     * @type {MatchListing}
     * @memberof GameRecordOutput
     */
    match?: MatchListing | null;
}

/**
 * Check if a given object implements the GameRecordOutput interface.
 */
export function instanceOfGameRecordOutput(value: object): value is GameRecordOutput {
    if (!('jsonS3Uri' in value) || value['jsonS3Uri'] === undefined) return false;
    if (!('gameTimestamp' in value) || value['gameTimestamp'] === undefined) return false;
    if (!('matchId' in value) || value['matchId'] === undefined) return false;
    if (!('replayFileUrl' in value) || value['replayFileUrl'] === undefined) return false;
    if (!('createdAt' in value) || value['createdAt'] === undefined) return false;
    if (!('gameDate' in value) || value['gameDate'] === undefined) return false;
    return true;
}

export function GameRecordOutputFromJSON(json: any): GameRecordOutput {
    return GameRecordOutputFromJSONTyped(json, false);
}

export function GameRecordOutputFromJSONTyped(json: any, ignoreDiscriminator: boolean): GameRecordOutput {
    if (json == null) {
        return json;
    }
    return {
        
        'jsonS3Uri': json['json_s3_uri'],
        'fileSizeBytes': json['file_size_bytes'] == null ? undefined : json['file_size_bytes'],
        'gameTimestamp': (new Date(json['game_timestamp'])),
        'matchId': json['match_id'],
        'replayFileUrl': json['replay_file_url'],
        'createdAt': (new Date(json['created_at'])),
        'gameDate': (new Date(json['game_date'])),
        'match': json['match'] == null ? undefined : MatchListingFromJSON(json['match']),
    };
}

export function GameRecordOutputToJSON(json: any): GameRecordOutput {
    return GameRecordOutputToJSONTyped(json, false);
}

export function GameRecordOutputToJSONTyped(value?: GameRecordOutput | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'json_s3_uri': value['jsonS3Uri'],
        'file_size_bytes': value['fileSizeBytes'],
        'game_timestamp': value['gameTimestamp'].toISOString(),
        'match_id': value['matchId'],
        'replay_file_url': value['replayFileUrl'],
        'created_at': value['createdAt'].toISOString(),
        'game_date': value['gameDate'].toISOString().substring(0,10),
        'match': MatchListingToJSON(value['match']),
    };
}